added header row functionality to layouts
header rows allow individual layout items to act as an entire row of content. it's useful in scenarios where a layout is being generated, but that layout has some sort of heading above it. instead of that heading acting as a normal layout item, it gets expanded to full-width and the rest of the items follow the normal order/positioning of rows/columns, almost acting as if that heading isn't a part of their layout but is instead a separate layout.